Bug fix for migration

+++ db/migrate/20130923092617_remove_start_time_from_speakers.rb
@@ -1,6 +1,11 @@
 class RemoveStartTimeFromSpeakers < ActiveRecord::Migration
   def change
-	remove_column :speakers, :start_time
-	remove_column :speakers, :end_time
+  	if  column_exists? :speakers, :start_time
+  		remove_column :speakers, :start_time
+  	end
+	
+	if column_exists? :speakers, :end_time
+		remove_column :speakers, :end_time
+	end
   end
 end
